♻️ [Refactor]: create study interface type 파일로 빼기
