+++ src/features/create-study/hooks/useStudyStepProtection.ts
@@ -3,7 +3,7 @@ import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect } from "react";
 
 import URL from "@/shared/constants/URL";
-import { StudyFormInterface } from "../form/CreateStudyForm";
+import { StudyFormInterface } from "../types/create-study.type";
 
 const useStudyStepProtection = (
   watch: (name?: string) => Partial<StudyFormInterface>
